+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # walqueue changelog
 
 - [FEATURE] Add support for headers in outgoing connections. (@mattdurham)
+- [FEATURE] Add support for proxies. (@mattdurham)